@@ -21,7 +21,7 @@
     "open": "ng serve --open",
     "build": "ng build --prod",
     "test": "ng test",
-    "test:single": "ng test --code-coverage --single-run",
+    "test:single": "ng test --code-coverage --watch false",
     "lint": "ng lint",
     "lint:fix": "ng lint --fix",
     "stylish": "ng lint --format stylish",
@@ -35,39 +35,40 @@
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
-    "@covalent/core": "^1.0.0",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.1.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.1.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@covalent/core": "2.0.0-beta.1",
     "@firebase/app": "~0.1.0",
-    "@ngrx/effects": "^5.2.0",
-    "@ngrx/entity": "^5.2.0",
-    "@ngrx/router-store": "^5.2.0",
-    "@ngrx/store": "^5.2.0",
-    "@ngrx/store-devtools": "^5.2.0",
-    "@ngx-formly/core": "^2.0.0-rc.55",
-    "@ngx-formly/material": "^2.0.0-rc.55",
-    "angularfire2": "^5.0.0-rc.6.0",
+    "@ngrx/effects": "6.0.1",
+    "@ngrx/entity": "6.0.1",
+    "@ngrx/router-store": "6.0.1",
+    "@ngrx/store": "6.0.1",
+    "@ngrx/store-devtools": "6.0.1",
+    "@ngx-formly/core": "4.3.0",
+    "@ngx-formly/material": "4.3.0",
+    "angularfire2": "^5.0.0-rc.10",
     "core-js": "^2.5.5",
-    "firebase": "^4.12.1",
+    "firebase": "5.0.4",
     "hammerjs": "^2.0.8",
-    "jasmine-marbles": "^0.2.0",
-    "rxjs": "^5.5.10",
+    "jasmine-marbles": "^0.3.1",
+    "rxjs": "^6.2.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.7.4",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/language-service": "^5.2.9",
+    "@angular-devkit/build-angular": "~0.6.5",
+    "@angular/cli": "^6.0.5",
+    "@angular/compiler-cli": "^6.0.3",
+    "@angular/language-service": "^6.0.3",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
@@ -90,7 +91,7 @@
     "rimraf": "^2.6.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "~2.6.2"
+    "typescript": "2.7.2"
   },
   "config": {
     "commitizen": {
