@@ -1,35 +1,28 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  Input,
-  OnInit
-} from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { ServerError, WarningTypeConfig } from '../../../shared';
+import { ChangeDetectionStrategy, Component, Input, OnInit } from '@angular/core';
+import { Observable } from 'rxjs';
+import { ServerError, toggleAnimation, WarningTypeConfig, WarningTypes, WarningTypesConfigList } from '../../../shared';
 
-import { toggleAnimation } from '../../../shared/animations/toggle.animation';
-import { WarningTypes, WarningTypesConfigList } from '../../../shared';
 import { ActivitySandbox } from '../../activity.sandbox';
 
 @Component({
   selector: 'app-alert',
   template: `
-	<div *ngIf="visible$ | async" @toggleAnimation>
-	  <td-message [label]="(alert$ | async)?.annexMessage"
-				  [sublabel]="(alert$ | async)?.message" [color]="type.color" [icon]="type.icon">
-		<button td-message-actions mat-icon-button (click)="close()">
-		  <mat-icon>cancel</mat-icon>
-		</button>
-	  </td-message>
-	</div>
+    <div *ngIf="visible$ | async" @toggleAnimation>
+      <td-message [label]="(alert$ | async)?.annexMessage"
+                  [sublabel]="(alert$ | async)?.message" [color]="type.color" [icon]="type.icon">
+        <button td-message-actions mat-icon-button (click)="close()">
+          <mat-icon>cancel</mat-icon>
+        </button>
+      </td-message>
+    </div>
   `,
-  animations: [toggleAnimation],
+  animations: [ toggleAnimation ],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class AlertComponent implements OnInit {
   visible$: Observable<boolean>;
   alert$: Observable<ServerError>;
-  @Input() type: WarningTypeConfig = WarningTypesConfigList[WarningTypes.Error];
+  @Input() type: WarningTypeConfig = WarningTypesConfigList[ WarningTypes.Error ];
 
   constructor(private sandbox: ActivitySandbox) {}
 
