@@ -1,8 +1,7 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { FormlyFieldConfig } from '@ngx-formly/core';
 
-import { Observable } from 'rxjs/Observable';
-import { Subscription } from 'rxjs/Subscription';
+import { Subscription } from 'rxjs';
 
 import { AuthService } from '../../../../services';
 import { User, WarningTypes, WarningTypesConfigList } from '../../../../shared';
@@ -16,19 +15,19 @@ import { ForgotSandbox } from '../../forgot.sandbox';
       <h3 title *ngIf="resetOk">Petició de canvi de clau rebuda</h3>
       <!-- form -->
       <app-card-content-form card-content-form *ngIf="!resetOk"
-      [fields]="fields"
-      [model]="model"
-      [submitTitle]="'Enviar'"
-      (submit)="onSubmit($event)">
-          <!-- login with form -->
-          <div otherAction class="mat-caption pad pad-top-none text-right">
-            <a [routerLink]="['../login']">Tornar al login</a>
-          </div>
-      </app-card-content-form>
-        <!-- login on reset ok -->
-      <div postSending *ngIf="resetOk" class="mat-caption pad pad-top-none text-right">
-          <h4 class="text-center">Consulta la teva bústia de correu</h4>
+                             [fields]="fields"
+                             [model]="model"
+                             [submitTitle]="'Enviar'"
+                             (submit)="onSubmit($event)">
+        <!-- login with form -->
+        <div otherAction class="mat-caption pad pad-top-none text-right">
           <a [routerLink]="['../login']">Tornar al login</a>
+        </div>
+      </app-card-content-form>
+      <!-- login on reset ok -->
+      <div postSending *ngIf="resetOk" class="mat-caption pad pad-top-none text-right">
+        <h4 class="text-center">Consulta la teva bústia de correu</h4>
+        <a [routerLink]="['../login']">Tornar al login</a>
       </div>
     </app-auth-container>
   `
@@ -37,7 +36,7 @@ export class ForgotComponent implements OnInit, OnDestroy {
   fields: Array<FormlyFieldConfig>;
   resetOk: boolean;
   model = {} as User;
-  type = WarningTypesConfigList[WarningTypes.Error];
+  type = WarningTypesConfigList[ WarningTypes.Error ];
   recoverSubscription: Subscription;
 
   constructor(
